% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_database.R
\name{calc_shrub_mass}
\alias{calc_shrub_mass}
\title{Calculate mass of shrub}
\usage{
calc_shrub_mass(volume)
}
\arguments{
\item{volume}{(numeric)}
}
\value{
mass of shrub (numeric)
calc_shrub_mass(3.5)
}
\description{
Calculate mass of shrub
}
